# swagger.yaml
openapi: 3.0.0
info:
  title: Teamitaka Practice API
  version: 1.0.0
  description: |
    Teamitaka 프로젝트 연습을 위한 API 명세서로서, 사용자 관리, 인증, 프로젝트 관리, 모집 공고, 프로필, 댓글, 스크랩, 검색, 대학 인증 등의 기능을 제공합니다.
    이 API는 Node.js와 Express.js를 기반으로 구현되었으며, Swagger UI를 통해 시각화됩니다.  
    주요 기능:
    - 사용자 등록 및 인증 (JWT 기반)
    - 관리자 기능 (관리자 계정 관리, 인증된 이메일 관리)
    - 모집 공고 생성/수정/삭제 및 댓글 관리
    - 프로젝트 생성/수정/삭제 및 멤버 관리
    - 프로필 생성/수정/삭제 및 이력서 조회
    - 스크랩 기능
    - 키워드 기반 검색
    - 대학 인증 (Univecert API 연동)
  contact:
    name: Teamitaka Support
    email: support@teamitaka.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:{port}
    description: 로컬 개발 서버
    variables:
      port:
        default: "3000"
        description: "서버 포트 (동적으로 설정 가능)"
  - url: https://api.teamitaka.com
    description: 프로덕션 서버
...
paths:
...
  /api/user:
    get:
      tags:
        - User
      summary: 사용자 목록 조회
      description: 모든 사용자 목록을 페이지네이션으로 반환합니다.
      operationId: getUsers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: 페이지 번호
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: 페이지당 항목 수
      responses:
        '200':
          description: 사용자 목록 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data:
                  users:
                    - user_id: 1
                      username: "johndoe"
                      email: "johndoe@example.com"
                      role: "MEMBER"
                      created_at: "2025-02-27T11:00:00.000Z"
                      updated_at: "2025-02-27T11:00:00.000Z"
                  total: 1
                  page: 1
                  limit: 10
                message: ""
        '401':
          description: 인증되지 않은 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "인증되지 않았습니다."
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "서버 오류 발생"
  /api/user/{id}:
    put:
      tags:
        - User
      summary: 사용자 정보 수정
      description: 특정 사용자의 정보를 수정합니다.
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 수정할 사용자 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "홍길동"
                  description: 새로운 사용자 이름
                email:
                  type: string
                  format: email
                  example: "hong@example.com"
                  description: 새로운 이메일 주소
              required:
                - name
                - email
      responses:
        '200':
          description: 사용자 정보 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data:
                  user:
                    user_id: 1
                    name: "홍길동"
                    email: "hong@example.com"
                    updated_at: "2025-02-27T11:00:00.000Z"
                message: "사용자 정보 수정 완료"
        '400':
          description: 필수 필드 누락
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "이름과 이메일은 필수입니다."
        '401':
          description: 인증되지 않은 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "인증되지 않았습니다."
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "사용자 수정 권한이 없습니다."
        '404':
          description: 사용자 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "사용자를 찾을 수 없습니다."
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "서버 오류 발생"
    delete:
      tags:
        - User
      summary: 사용자 삭제
      description: 특정 사용자를 삭제합니다.
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 삭제할 사용자 ID
      responses:
        '200':
          description: 사용자 삭제 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data: {}
                message: "사용자 삭제 완료"
        '401':
          description: 인증되지 않은 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "인증되지 않았습니다."
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "사용자 삭제 권한이 없습니다."
        '404':
          description: 사용자 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "사용자를 찾을 수 없습니다."
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "서버 오류 발생"

  # 모집 공고 관련 엔드포인트
...
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 요청 성공 여부
        data:
          type: object
          description: 반환 데이터
        message:
          type: string
          description: 추가 메시지
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
          description: 요청 실패 여부
        message:
          type: string
          description: 오류 메시지
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT 기반 인증 토큰 (Bearer 형식)
...